rename global module file from all to index
